Trait abcf::module::Application

pub trait Application: Send + Sync + Sized + ModuleStorage + Dependence {
    type Transaction: Default + Send + Sync;
    fn check_tx<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self, 
        _context: &'life1 mut TxnContext<'life2, Self>, 
        _req: &'life3 RequestCheckTx<Self::Transaction>
    ) -> Pin<Box<dyn Future<Output = Result<ResponseCheckTx>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn begin_block<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self, 
        _context: &'life1 mut AppContext<'life2, Self>, 
        _req: &'life3 RequestBeginBlock
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn deliver_tx<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self, 
        _context: &'life1 mut TxnContext<'life2, Self>, 
        _req: &'life3 RequestDeliverTx<Self::Transaction>
    ) -> Pin<Box<dyn Future<Output = Result<ResponseDeliverTx>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn end_block<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 mut self, 
        _context: &'life1 mut AppContext<'life2, Self>, 
        _req: &'life3 RequestEndBlock
    ) -> Pin<Box<dyn Future<Output = ResponseEndBlock> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait,
    { ... }
}

This trait define module’s main blockchain logic.

Associated Types

type Transaction: Default + Send + Sync

Provided methods

fn check_tx<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 mut self,
_context: &'life1 mut TxnContext<'life2, Self>,
_req: &'life3 RequestCheckTx<Self::Transaction>
) -> Pin<Box<dyn Future<Output = Result<ResponseCheckTx>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Define how to check transaction.

In this function, do some lightweight check for transaction, for example: check signature, check balance and so on. This method will be called at external user or another node.

fn begin_block<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 mut self,
_context: &'life1 mut AppContext<'life2, Self>,
_req: &'life3 RequestBeginBlock
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Begin block.

fn deliver_tx<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 mut self,
_context: &'life1 mut TxnContext<'life2, Self>,
_req: &'life3 RequestDeliverTx<Self::Transaction>
) -> Pin<Box<dyn Future<Output = Result<ResponseDeliverTx>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Execute transaction on state.

fn end_block<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 mut self,
_context: &'life1 mut AppContext<'life2, Self>,
_req: &'life3 RequestEndBlock
) -> Pin<Box<dyn Future<Output = ResponseEndBlock> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

End Block.